refactor(store)!: introduce Store<S> wrapper, modularize nexus-store
- Add Store<S> as Arc-wrapped shared handle to any RawEventStore
- Store::repository(codec, upcaster) creates typed EventStore facades
- Store::zero_copy_repository(codec, upcaster) for borrowing codecs
- Split EventStore and ZeroCopyEventStore into separate modules
- Reorganize nexus-store from flat files into codec/, envelope/,
  store/, and upcasting/ directories
- Migrate all tests and examples to new Store::repository() API
- Update CLAUDE.md architecture documentation
