+++ crates/nexus-fjall/benches/fjall_benchmarks.rs
@@ -15,8 +15,8 @@ use nexus_fjall::encoding::{
 };
 use nexus_store::PendingEnvelope;
 use nexus_store::envelope::pending_envelope;
-use nexus_store::raw::RawEventStore;
-use nexus_store::stream::EventStream;
+use nexus_store::store::EventStream;
+use nexus_store::store::RawEventStore;
 use std::fmt;
 use tempfile::TempDir;
 use tokio::runtime::Runtime;
