+++ crates/nexus-fjall/src/stream.rs
@@ -3,7 +3,7 @@ use crate::error::FjallError;
 use fjall::Slice;
 use nexus::Version;
 use nexus_store::PersistedEnvelope;
-use nexus_store::stream::EventStream;
+use nexus_store::store::EventStream;
 
 /// Lending cursor over fjall event rows.
 ///
