+++ crates/nexus-fjall/tests/property_tests.rs
@@ -52,8 +52,8 @@ use nexus_fjall::encoding::{
 use nexus_store::PendingEnvelope;
 use nexus_store::envelope::pending_envelope;
 use nexus_store::error::AppendError;
-use nexus_store::raw::RawEventStore;
-use nexus_store::stream::EventStream;
+use nexus_store::store::EventStream;
+use nexus_store::store::RawEventStore;
 
 use proptest::prelude::*;
