+++ crates/nexus-fjall/tests/state_machine_tests.rs
@@ -26,8 +26,8 @@ use nexus_fjall::FjallStore;
 use nexus_store::PendingEnvelope;
 use nexus_store::envelope::pending_envelope;
 use nexus_store::error::AppendError;
-use nexus_store::raw::RawEventStore;
-use nexus_store::stream::EventStream;
+use nexus_store::store::EventStream;
+use nexus_store::store::RawEventStore;
 use proptest::prelude::*;
 use proptest::strategy::BoxedStrategy;
 use proptest_state_machine::{ReferenceStateMachine, StateMachineTest, prop_state_machine};
