+++ crates/nexus-store/src/codec/borrowing.rs
