+++ crates/nexus-store/src/codec/mod.rs
@@ -1,0 +1,5 @@
+mod borrowing;
+mod owning;
+
+pub use borrowing::BorrowingCodec;
+pub use owning::Codec;
