+++ crates/nexus-store/src/codec/owning.rs
