+++ crates/nexus-store/src/lib.rs
@@ -1,26 +1,16 @@
-pub mod borrowing_codec;
 pub mod codec;
 pub mod envelope;
 pub mod error;
-pub mod event_store;
-pub mod morsel;
-pub mod raw;
-pub mod repository;
-pub mod stream;
+pub mod store;
 pub mod stream_label;
 #[cfg(feature = "testing")]
 pub mod testing;
-pub mod upcaster;
+pub mod upcasting;
 
-pub use borrowing_codec::BorrowingCodec;
-pub use codec::Codec;
+pub use codec::{BorrowingCodec, Codec};
 pub use envelope::{PendingEnvelope, PersistedEnvelope, pending_envelope};
 pub use error::{AppendError, InvalidSchemaVersion, StoreError, UpcastError};
-pub use event_store::{EventStore, ZeroCopyEventStore};
-pub use morsel::EventMorsel;
 pub use nexus::Version;
-pub use raw::RawEventStore;
-pub use repository::Repository;
-pub use stream::EventStream;
+pub use store::{EventStore, EventStream, RawEventStore, Repository, Store, ZeroCopyEventStore};
 pub use stream_label::{StreamLabel, ToStreamLabel};
-pub use upcaster::Upcaster;
+pub use upcasting::{EventMorsel, Upcaster};
