+++ crates/nexus-store/src/store/mod.rs
@@ -1,0 +1,17 @@
+mod event_store;
+mod raw;
+mod repository;
+#[allow(
+    clippy::module_inception,
+    reason = "Store<S> is the primary public type of the store module"
+)]
+mod store;
+mod stream;
+mod zero_copy_event_store;
+
+pub use event_store::EventStore;
+pub use raw::RawEventStore;
+pub use repository::Repository;
+pub use store::Store;
+pub use stream::EventStream;
+pub use zero_copy_event_store::ZeroCopyEventStore;
