+++ crates/nexus-store/src/store/stream.rs
