+++ crates/nexus-store/src/upcasting/mod.rs
@@ -1,0 +1,5 @@
+mod morsel;
+mod upcaster;
+
+pub use morsel::EventMorsel;
+pub use upcaster::Upcaster;
