+++ crates/nexus-store/tests/bug_hunt_tests.rs
@@ -32,8 +32,8 @@ use nexus_store::ToStreamLabel;
 use nexus_store::envelope::{PendingEnvelope, PersistedEnvelope};
 use nexus_store::error::StoreError;
 use nexus_store::pending_envelope;
-use nexus_store::raw::RawEventStore;
-use nexus_store::stream::EventStream;
+use nexus_store::store::EventStream;
+use nexus_store::store::RawEventStore;
 use std::collections::HashMap;
 use std::fmt;
 use tokio::sync::Mutex;
