+++ crates/nexus-store/tests/repository_qa_tests.rs
@@ -41,13 +41,13 @@ use std::sync::atomic::{AtomicBool, AtomicUsize, Ordering};
 
 use nexus::*;
 use nexus_store::BorrowingCodec;
+use nexus_store::Store;
 use nexus_store::codec::Codec;
 use nexus_store::error::StoreError;
-use nexus_store::event_store::{EventStore, ZeroCopyEventStore};
-use nexus_store::morsel::EventMorsel;
-use nexus_store::raw::RawEventStore;
-use nexus_store::repository::Repository;
+use nexus_store::store::RawEventStore;
+use nexus_store::store::Repository;
 use nexus_store::testing::InMemoryStore;
+use nexus_store::upcasting::EventMorsel;
 use nexus_store::{UpcastError, Upcaster};
 use proptest::prelude::*;
 
@@ -440,7 +440,8 @@ impl Upcaster for DeltaDoublingUpcaster {
 #[tokio::test]
 async fn d1_encode_failure_mid_batch_returns_codec_error() {
     // Encode succeeds for 1st event, fails for 2nd
-    let es = EventStore::new(InMemoryStore::new(), FailAfterNEncodesCodec::new(1));
+    let store = Store::new(InMemoryStore::new());
+    let es = store.repository(FailAfterNEncodesCodec::new(1), ());
     let mut agg = AggregateRoot::<CounterAggregate>::new(CounterId(1));
 
     let result = es
@@ -463,7 +464,8 @@ async fn d1_encode_failure_mid_batch_returns_codec_error() {
 #[tokio::test]
 async fn d1_encode_failure_first_event_returns_codec_error() {
     // All encodes fail — no events should reach the store
-    let es = EventStore::new(InMemoryStore::new(), FailAfterNEncodesCodec::new(0));
+    let store = Store::new(InMemoryStore::new());
+    let es = store.repository(FailAfterNEncodesCodec::new(0), ());
     let mut agg = AggregateRoot::<CounterAggregate>::new(CounterId(1));
 
     let result = es.save(&mut agg, &[CounterEvent::Incremented]).await;
@@ -474,7 +476,8 @@ async fn d1_encode_failure_first_event_returns_codec_error() {
 async fn d1_decode_failure_mid_stream_returns_codec_error() {
     // FailAfterNDecodesCodec: encode always succeeds, decode fails after N calls.
     // Save 3 events (encodes fine), then load (2nd decode fails).
-    let es = EventStore::new(InMemoryStore::new(), FailAfterNDecodesCodec::new(1));
+    let store = Store::new(InMemoryStore::new());
+    let es = store.repository(FailAfterNDecodesCodec::new(1), ());
     let mut agg = AggregateRoot::<CounterAggregate>::new(CounterId(1));
     es.save(
         &mut agg,
@@ -499,7 +502,8 @@ async fn d1_decode_failure_mid_stream_returns_codec_error() {
 #[tokio::test]
 async fn d1_decode_failure_first_event_returns_codec_error() {
     // Decode fails on the very first event
-    let es = EventStore::new(InMemoryStore::new(), FailAfterNDecodesCodec::new(0));
+    let store = Store::new(InMemoryStore::new());
+    let es = store.repository(FailAfterNDecodesCodec::new(0), ());
     let mut agg = AggregateRoot::<CounterAggregate>::new(CounterId(1));
     es.save(&mut agg, &[CounterEvent::Incremented])
         .await
@@ -518,7 +522,8 @@ async fn d1_decode_failure_first_event_returns_codec_error() {
 
 #[tokio::test]
 async fn d2_failed_encode_must_not_advance_version() {
-    let es = EventStore::new(InMemoryStore::new(), FailAfterNEncodesCodec::new(0));
+    let store = Store::new(InMemoryStore::new());
+    let es = store.repository(FailAfterNEncodesCodec::new(0), ());
     let mut agg = AggregateRoot::<CounterAggregate>::new(CounterId(1));
 
     assert_eq!(agg.version(), None);
@@ -546,7 +551,8 @@ async fn d2_aggregate_must_be_retryable_after_failed_save() {
     let codec = ToggleableCodec {
         fail_flag: fail_flag.clone(),
     };
-    let es = EventStore::new(InMemoryStore::new(), codec);
+    let store = Store::new(InMemoryStore::new());
+    let es = store.repository(codec, ());
     let mut agg = AggregateRoot::<CounterAggregate>::new(CounterId(1));
 
     let events = [CounterEvent::Incremented, CounterEvent::Incremented];
@@ -578,7 +584,8 @@ async fn d2_aggregate_must_be_retryable_after_failed_save() {
 
 #[tokio::test(flavor = "multi_thread", worker_threads = 2)]
 async fn d3_concurrent_saves_one_wins_one_conflicts() {
-    let es = Arc::new(EventStore::new(InMemoryStore::new(), SimpleCodec));
+    let store = Store::new(InMemoryStore::new());
+    let es = Arc::new(store.repository(SimpleCodec, ()));
 
     // Setup: one event in the stream
     let mut setup = AggregateRoot::<CounterAggregate>::new(CounterId(1));
@@ -632,7 +639,8 @@ async fn d3_concurrent_saves_one_wins_one_conflicts() {
 
 #[tokio::test(flavor = "multi_thread", worker_threads = 2)]
 async fn d3_concurrent_loads_both_succeed() {
-    let es = Arc::new(EventStore::new(InMemoryStore::new(), SimpleCodec));
+    let store = Store::new(InMemoryStore::new());
+    let es = Arc::new(store.repository(SimpleCodec, ()));
 
     let mut agg = AggregateRoot::<CounterAggregate>::new(CounterId(1));
     es.save(&mut agg, &[CounterEvent::Set(42)]).await.unwrap();
@@ -659,7 +667,8 @@ async fn d3_concurrent_loads_both_succeed() {
 
 #[tokio::test]
 async fn d3_cross_stream_concurrent_saves_both_succeed() {
-    let es = Arc::new(EventStore::new(InMemoryStore::new(), SimpleCodec));
+    let store = Store::new(InMemoryStore::new());
+    let es = Arc::new(store.repository(SimpleCodec, ()));
 
     let handle_a = tokio::spawn({
         let es = es.clone();
@@ -692,7 +701,8 @@ async fn d3_cross_stream_concurrent_saves_both_succeed() {
 
 #[tokio::test]
 async fn d4_single_upcaster_transforms_on_load() {
-    let es = EventStore::with_upcaster(InMemoryStore::new(), SimpleCodec, IncrementedV1ToV2);
+    let store = Store::new(InMemoryStore::new());
+    let es = store.repository(SimpleCodec, IncrementedV1ToV2);
 
     let mut agg = AggregateRoot::<CounterAggregate>::new(CounterId(1));
     es.save(&mut agg, &[CounterEvent::Set(42)]).await.unwrap();
@@ -705,7 +715,8 @@ async fn d4_single_upcaster_transforms_on_load() {
 
 #[tokio::test]
 async fn d4_chained_upcasters_v1_to_v3() {
-    let es = EventStore::with_upcaster(InMemoryStore::new(), SimpleCodec, IncrementedV1ToV3);
+    let store = Store::new(InMemoryStore::new());
+    let es = store.repository(SimpleCodec, IncrementedV1ToV3);
     // Flow: v1 → v2 → v3 (upcaster handles both steps internally)
 
     let mut agg = AggregateRoot::<CounterAggregate>::new(CounterId(1));
@@ -752,8 +763,8 @@ async fn d4_zero_copy_event_store_with_payload_mutating_upcaster() {
         .unwrap();
 
     // Load via ZeroCopyEventStore with upcaster — delta=5 doubled to 10
-    let es =
-        ZeroCopyEventStore::with_upcaster(raw_store, DeltaBorrowingCodec, DeltaDoublingUpcaster);
+    let store = Store::new(raw_store);
+    let es = store.zero_copy_repository(DeltaBorrowingCodec, DeltaDoublingUpcaster);
     let loaded: AggregateRoot<DeltaAggregate> = es.load(CounterId(1)).await.unwrap();
     assert_eq!(
         loaded.state().total,
@@ -781,8 +792,8 @@ async fn d4_upcaster_must_not_double_apply_to_new_events() {
         .await
         .unwrap();
 
-    let es =
-        ZeroCopyEventStore::with_upcaster(raw_store, DeltaBorrowingCodec, DeltaDoublingUpcaster);
+    let store = Store::new(raw_store);
+    let es = store.zero_copy_repository(DeltaBorrowingCodec, DeltaDoublingUpcaster);
 
     // Load: legacy delta=5 upcasted to 10. State total=10.
     let mut loaded: AggregateRoot<DeltaAggregate> = es.load(CounterId(1)).await.unwrap();
@@ -810,7 +821,8 @@ async fn d4_upcaster_must_not_double_apply_to_new_events() {
 
 #[tokio::test]
 async fn d5_first_save_uses_version_initial_as_expected() {
-    let es = EventStore::new(InMemoryStore::new(), SimpleCodec);
+    let store = Store::new(InMemoryStore::new());
+    let es = store.repository(SimpleCodec, ());
 
     let mut agg = AggregateRoot::<CounterAggregate>::new(CounterId(1));
     es.save(&mut agg, &[CounterEvent::Incremented])
@@ -825,7 +837,8 @@ async fn d5_first_save_uses_version_initial_as_expected() {
 
 #[tokio::test]
 async fn d5_version_consistency_through_save_load_cycles() {
-    let es = EventStore::new(InMemoryStore::new(), SimpleCodec);
+    let store = Store::new(InMemoryStore::new());
+    let es = store.repository(SimpleCodec, ());
 
     let mut agg = AggregateRoot::<CounterAggregate>::new(CounterId(1));
 
@@ -855,7 +868,8 @@ async fn d5_version_consistency_through_save_load_cycles() {
 
 #[tokio::test]
 async fn d5_batch_version_tracking() {
-    let es = EventStore::new(InMemoryStore::new(), SimpleCodec);
+    let store = Store::new(InMemoryStore::new());
+    let es = store.repository(SimpleCodec, ());
     let mut agg = AggregateRoot::<CounterAggregate>::new(CounterId(1));
 
     let events = [
@@ -878,7 +892,8 @@ async fn d5_batch_version_tracking() {
 
 #[tokio::test]
 async fn d6_full_lifecycle_fresh_save_load_modify_save_load() {
-    let es = EventStore::new(InMemoryStore::new(), SimpleCodec);
+    let store = Store::new(InMemoryStore::new());
+    let es = store.repository(SimpleCodec, ());
 
     // 1. Fresh aggregate
     let mut agg = AggregateRoot::<CounterAggregate>::new(CounterId(1));
@@ -911,7 +926,8 @@ async fn d6_full_lifecycle_fresh_save_load_modify_save_load() {
 
 #[tokio::test]
 async fn d6_ten_round_modify_save_load_cycles() {
-    let es = EventStore::new(InMemoryStore::new(), SimpleCodec);
+    let store = Store::new(InMemoryStore::new());
+    let es = store.repository(SimpleCodec, ());
 
     let mut agg = AggregateRoot::<CounterAggregate>::new(CounterId(1));
 
@@ -931,7 +947,8 @@ async fn d6_ten_round_modify_save_load_cycles() {
 
 #[tokio::test]
 async fn d6_state_determinism_same_events_same_state() {
-    let es = EventStore::new(InMemoryStore::new(), SimpleCodec);
+    let store = Store::new(InMemoryStore::new());
+    let es = store.repository(SimpleCodec, ());
 
     let mut agg = AggregateRoot::<CounterAggregate>::new(CounterId(1));
     es.save(
@@ -957,7 +974,8 @@ async fn d6_state_determinism_same_events_same_state() {
 
 #[tokio::test]
 async fn d6_load_after_save_has_no_pending_state() {
-    let es = EventStore::new(InMemoryStore::new(), SimpleCodec);
+    let store = Store::new(InMemoryStore::new());
+    let es = store.repository(SimpleCodec, ());
 
     let mut agg = AggregateRoot::<CounterAggregate>::new(CounterId(1));
     es.save(&mut agg, &[CounterEvent::Incremented])
@@ -975,7 +993,8 @@ async fn d6_load_after_save_has_no_pending_state() {
 
 #[tokio::test]
 async fn d7_save_empty_is_noop() {
-    let es = EventStore::new(InMemoryStore::new(), SimpleCodec);
+    let store = Store::new(InMemoryStore::new());
+    let es = store.repository(SimpleCodec, ());
     let mut agg = AggregateRoot::<CounterAggregate>::new(CounterId(1));
 
     // Save with empty events — should be a no-op
@@ -985,7 +1004,8 @@ async fn d7_save_empty_is_noop() {
 
 #[tokio::test]
 async fn d7_save_advances_version() {
-    let es = EventStore::new(InMemoryStore::new(), SimpleCodec);
+    let store = Store::new(InMemoryStore::new());
+    let es = store.repository(SimpleCodec, ());
     let mut agg = AggregateRoot::<CounterAggregate>::new(CounterId(1));
 
     es.save(
@@ -999,7 +1019,8 @@ async fn d7_save_advances_version() {
 
 #[tokio::test]
 async fn d7_load_nonexistent_stream_returns_fresh_aggregate() {
-    let es = EventStore::new(InMemoryStore::new(), SimpleCodec);
+    let store = Store::new(InMemoryStore::new());
+    let es = store.repository(SimpleCodec, ());
     let loaded: AggregateRoot<CounterAggregate> = es.load(CounterId(1)).await.unwrap();
 
     assert_eq!(loaded.version(), None);
@@ -1008,7 +1029,8 @@ async fn d7_load_nonexistent_stream_returns_fresh_aggregate() {
 
 #[tokio::test]
 async fn d7_repository_preserves_event_ordering() {
-    let es = EventStore::new(InMemoryStore::new(), SimpleCodec);
+    let store = Store::new(InMemoryStore::new());
+    let es = store.repository(SimpleCodec, ());
     let mut agg = AggregateRoot::<CounterAggregate>::new(CounterId(1));
 
     es.save(
@@ -1037,7 +1059,8 @@ async fn d7_repository_preserves_event_ordering() {
 
 #[tokio::test]
 async fn d8_codec_encode_error_is_store_error_codec() {
-    let es = EventStore::new(InMemoryStore::new(), FailAfterNEncodesCodec::new(0));
+    let store = Store::new(InMemoryStore::new());
+    let es = store.repository(FailAfterNEncodesCodec::new(0), ());
     let mut agg = AggregateRoot::<CounterAggregate>::new(CounterId(1));
 
     match es.save(&mut agg, &[CounterEvent::Incremented]).await {
@@ -1053,7 +1076,8 @@ async fn d8_codec_encode_error_is_store_error_codec() {
 
 #[tokio::test]
 async fn d8_codec_decode_error_is_store_error_codec() {
-    let es = EventStore::new(InMemoryStore::new(), FailAfterNDecodesCodec::new(0));
+    let store = Store::new(InMemoryStore::new());
+    let es = store.repository(FailAfterNDecodesCodec::new(0), ());
     let mut agg = AggregateRoot::<CounterAggregate>::new(CounterId(1));
     es.save(&mut agg, &[CounterEvent::Incremented])
         .await
@@ -1070,7 +1094,8 @@ async fn d8_codec_decode_error_is_store_error_codec() {
 
 #[tokio::test]
 async fn d8_concurrency_conflict_should_be_store_error_conflict_not_adapter() {
-    let es = EventStore::new(InMemoryStore::new(), SimpleCodec);
+    let store = Store::new(InMemoryStore::new());
+    let es = store.repository(SimpleCodec, ());
 
     let mut agg1 = AggregateRoot::<CounterAggregate>::new(CounterId(1));
     es.save(&mut agg1, &[CounterEvent::Incremented])
@@ -1096,7 +1121,8 @@ async fn d8_concurrency_conflict_should_be_store_error_conflict_not_adapter() {
 
 #[tokio::test]
 async fn d8_rehydration_limit_exceeded_is_store_error_kernel() {
-    let es = EventStore::new(InMemoryStore::new(), SimpleCodec);
+    let store = Store::new(InMemoryStore::new());
+    let es = store.repository(SimpleCodec, ());
 
     // Save 6 events (3 + 3) using TinyAggregate (MAX_REHYDRATION_EVENTS=5)
     let mut agg = AggregateRoot::<TinyAggregate>::new(CounterId(1));
@@ -1152,7 +1178,8 @@ proptest! {
     ) {
         let rt = tokio::runtime::Runtime::new().unwrap();
         rt.block_on(async {
-            let es = EventStore::new(InMemoryStore::new(), SimpleCodec);
+            let store = Store::new(InMemoryStore::new());
+            let es = store.repository(SimpleCodec, ());
             let mut agg = AggregateRoot::<CounterAggregate>::new(CounterId(1));
 
             // Compute expected state
@@ -1178,7 +1205,8 @@ proptest! {
     ) {
         let rt = tokio::runtime::Runtime::new().unwrap();
         rt.block_on(async {
-            let es = EventStore::new(InMemoryStore::new(), SimpleCodec);
+            let store = Store::new(InMemoryStore::new());
+            let es = store.repository(SimpleCodec, ());
 
             let expected_value = events.iter().fold(0i64, |acc, e| match e {
                 CounterEvent::Incremented => acc + 1,
@@ -1202,7 +1230,8 @@ proptest! {
     ) {
         let rt = tokio::runtime::Runtime::new().unwrap();
         rt.block_on(async {
-            let es = EventStore::new(InMemoryStore::new(), SimpleCodec);
+            let store = Store::new(InMemoryStore::new());
+            let es = store.repository(SimpleCodec, ());
 
             let mut agg = AggregateRoot::<CounterAggregate>::new(CounterId(1));
             es.save(&mut agg, &events).await.unwrap();
@@ -1226,7 +1255,8 @@ proptest! {
 
 #[tokio::test]
 async fn d10_large_batch_500_events_save_and_load() {
-    let es = EventStore::new(InMemoryStore::new(), SimpleCodec);
+    let store = Store::new(InMemoryStore::new());
+    let es = store.repository(SimpleCodec, ());
     let mut agg = AggregateRoot::<CounterAggregate>::new(CounterId(1));
 
     let events: Vec<CounterEvent> = (0..500).map(|_| CounterEvent::Incremented).collect();
@@ -1240,7 +1270,8 @@ async fn d10_large_batch_500_events_save_and_load() {
 #[tokio::test]
 async fn d10_max_rehydration_events_boundary() {
     // TinyAggregate: MAX_REHYDRATION_EVENTS = 5
-    let es = EventStore::new(InMemoryStore::new(), SimpleCodec);
+    let store = Store::new(InMemoryStore::new());
+    let es = store.repository(SimpleCodec, ());
 
     // Save exactly 5 events (3 + 2 via reload)
     let mut agg = AggregateRoot::<TinyAggregate>::new(CounterId(1));
@@ -1279,7 +1310,7 @@ async fn d10_max_rehydration_events_boundary() {
 #[tokio::test]
 async fn d11_schema_version_always_one() {
     use nexus_store::pending_envelope;
-    use nexus_store::stream::EventStream;
+    use nexus_store::store::EventStream;
 
     let store = InMemoryStore::new();
 
@@ -1316,7 +1347,8 @@ async fn d11_schema_version_always_one() {
 
 #[tokio::test]
 async fn d11_save_and_load_with_typed_id() {
-    let es = EventStore::new(InMemoryStore::new(), SimpleCodec);
+    let store = Store::new(InMemoryStore::new());
+    let es = store.repository(SimpleCodec, ());
     let mut agg = AggregateRoot::<CounterAggregate>::new(CounterId(1));
     let result = es.save(&mut agg, &[CounterEvent::Incremented]).await;
     assert!(result.is_ok(), "save with typed id should be accepted");
@@ -1327,7 +1359,8 @@ async fn d11_save_and_load_with_typed_id() {
 
 #[tokio::test]
 async fn d11_stream_isolation_different_streams_independent() {
-    let es = EventStore::new(InMemoryStore::new(), SimpleCodec);
+    let store = Store::new(InMemoryStore::new());
+    let es = store.repository(SimpleCodec, ());
 
     let mut agg_a = AggregateRoot::<CounterAggregate>::new(CounterId(1));
     es.save(&mut agg_a, &[CounterEvent::Set(100)])
@@ -1353,7 +1386,8 @@ async fn d11_stream_isolation_different_streams_independent() {
 
 #[tokio::test]
 async fn d11_append_empty_batch_is_noop() {
-    let es = EventStore::new(InMemoryStore::new(), SimpleCodec);
+    let store = Store::new(InMemoryStore::new());
+    let es = store.repository(SimpleCodec, ());
 
     let mut agg = AggregateRoot::<CounterAggregate>::new(CounterId(1));
     es.save(&mut agg, &[]).await.unwrap();
@@ -1364,7 +1398,8 @@ async fn d11_append_empty_batch_is_noop() {
 
 #[tokio::test]
 async fn d11_multiple_event_types_in_single_stream() {
-    let es = EventStore::new(InMemoryStore::new(), SimpleCodec);
+    let store = Store::new(InMemoryStore::new());
+    let es = store.repository(SimpleCodec, ());
     let mut agg = AggregateRoot::<CounterAggregate>::new(CounterId(1));
 
     es.save(
