+++ crates/nexus-store/tests/security_tests.rs
@@ -30,8 +30,8 @@ use nexus::{Id, Version};
 use nexus_store::ToStreamLabel;
 use nexus_store::error::StoreError;
 use nexus_store::pending_envelope;
-use nexus_store::raw::RawEventStore;
-use nexus_store::stream::EventStream;
+use nexus_store::store::EventStream;
+use nexus_store::store::RawEventStore;
 use nexus_store::testing::InMemoryStore;
 
 #[derive(Debug, Clone, Hash, PartialEq, Eq)]
