+++ examples/store-and-kernel/src/main.rs
@@ -21,8 +21,8 @@
 #![allow(clippy::shadow_unrelated, reason = "example shadows for readability")]
 
 use nexus::*;
-use nexus_store::raw::RawEventStore;
-use nexus_store::stream::EventStream;
+use nexus_store::store::EventStream;
+use nexus_store::store::RawEventStore;
 use nexus_store::testing::InMemoryStore;
 use nexus_store::{Codec, pending_envelope};
 use serde::{Deserialize, Serialize};
