+++ examples/store-inmemory/src/main.rs
@@ -23,11 +23,11 @@
 
 use nexus::Id;
 use nexus::Version;
-use nexus_store::morsel::EventMorsel;
-use nexus_store::raw::RawEventStore;
-use nexus_store::stream::EventStream;
+use nexus_store::store::EventStream;
+use nexus_store::store::RawEventStore;
 use nexus_store::testing::InMemoryStore;
-use nexus_store::upcaster::Upcaster;
+use nexus_store::upcasting::EventMorsel;
+use nexus_store::upcasting::Upcaster;
 use nexus_store::{Codec, pending_envelope};
 use serde::{Deserialize, Serialize};
 use std::fmt;
